# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ecrossflow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email/username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout and invalidate session
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/verify-referral:
    get:
      operationId: verifyReferral
      tags: [auth]
      summary: Verify a referral code and get referrer info
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Referral code info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralVerifyResponse"
        "404":
          description: Code not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/check-username:
    get:
      operationId: checkUsername
      tags: [auth]
      summary: Check if username is available
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Username availability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsernameCheckResponse"

  # USER ENDPOINTS
...
  /wallet:
    get:
      operationId: getWallet
      tags: [wallet]
      summary: Get wallet balance
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Wallet balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBalance"
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get platform statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Platform stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: Get all users (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
  /admin/users/{id}/suspend:
    put:
      operationId: suspendUser
      tags: [admin]
      summary: Suspend a user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/users/{id}/activate:
    put:
      operationId: activateUser
      tags: [admin]
      summary: Activate a user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/users/{id}/adjust-balance:
    post:
      operationId: adjustUserBalance
      tags: [admin]
      summary: Manually adjust user balance
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustBalanceRequest"
      responses:
        "200":
          description: Balance adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/deposits/pending:
    get:
      operationId: getPendingDeposits
      tags: [admin]
      summary: Get pending manual deposits (Moncash/Natcash)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Pending deposits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDepositListResponse"
  /admin/deposits/{id}/approve:
    put:
      operationId: approveDeposit
      tags: [admin]
      summary: Approve a pending deposit
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deposit approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/deposits/{id}/reject:
    put:
      operationId: rejectDeposit
      tags: [admin]
      summary: Reject a pending deposit
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectDepositRequest"
      responses:
        "200":
          description: Deposit rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # AUTH SCHEMAS
    RegisterRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
          minLength: 3
          maxLength: 20
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          minLength: 8
        referralCode:
          type: string
        preferredLanguage:
          type: string
          default: "fr"
      required:
        - firstName
        - lastName
        - username
        - email
        - password
        - referralCode
    LoginRequest:
      type: object
      properties:
        emailOrUsername:
          type: string
        password:
          type: string
        rememberMe:
          type: boolean
          default: false
      required:
        - emailOrUsername
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    ReferralVerifyResponse:
      type: object
      properties:
        valid:
          type: boolean
        referrerName:
          type: string
        referrerUsername:
          type: string
      required:
        - valid
    UsernameCheckResponse:
      type: object
      properties:
        available:
          type: boolean
      required:
        - available

    # USER SCHEMAS
    UserProfile:
      type: object
      properties:
        id:
          type: string
        accountNumber:
          type: integer
          nullable: true
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        referralCode:
          type: string
        status:
          type: string
          enum: [PENDING, ACTIVE, SUSPENDED]
        role:
          type: string
          enum: [USER, ADMIN]
        preferredLanguage:
          type: string
        preferredCurrency:
          type: string
        preferredTheme:
          type: string
        currentBoard:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - username
        - email
        - referralCode
        - status
        - role
        - preferredLanguage
        - preferredCurrency
        - preferredTheme
        - createdAt
...
    WalletBalance:
      type: object
      properties:
        balanceUsd:
          type: number
        balancePending:
          type: number
        balanceReserved:
          type: number
        totalBalance:
          type: number
      required:
        - balanceUsd
        - balancePending
        - balanceReserved
        - totalBalance
...
    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        activeUsers:
          type: integer
        pendingUsers:
          type: integer
        totalVolume24h:
          type: number
        totalVolume7d:
          type: number
        activeBoards:
          type: integer
        pendingDeposits:
          type: integer
        pendingWithdrawals:
          type: integer
        totalPlatformRevenue:
          type: number
      required:
        - totalUsers
        - activeUsers
        - pendingUsers
        - totalVolume24h
        - totalVolume7d
        - activeBoards
        - pendingDeposits
        - pendingWithdrawals
        - totalPlatformRevenue
    AdminUserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - users
        - total
        - page
        - totalPages
    AdminUser:
      type: object
      properties:
        id:
          type: string
        accountNumber:
          type: integer
          nullable: true
        username:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        status:
          type: string
        role:
          type: string
        currentBoard:
          type: string
          nullable: true
        walletBalance:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
        - firstName
        - lastName
        - status
        - role
        - walletBalance
        - createdAt
    AdjustBalanceRequest:
      type: object
      properties:
        amount:
          type: number
        note:
          type: string
      required:
        - amount
        - note
    AdminDepositListResponse:
      type: object
      properties:
        deposits:
          type: array
          items:
            $ref: "#/components/schemas/AdminDeposit"
        total:
          type: integer
      required:
        - deposits
        - total
    AdminDeposit:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        username:
          type: string
        amount:
          type: number
        currency:
          type: string
        amountHtg:
          type: number
          nullable: true
        paymentMethod:
          type: string
        reference:
          type: string
          nullable: true
        screenshotUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - username
        - amount
        - currency
        - paymentMethod
        - createdAt
    RejectDepositRequest:
      type: object
      properties:
        reason:
          type: string
      required:
        - reason