# lib/api-zod/src/generated/types/loginRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ecrossflow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LoginRequest {
  emailOrUsername: string;
  password: string;
  rememberMe?: boolean;
}